1. remove all code from the original project 2. update package.json
