--- src/bindings.d.ts
@@ -1,8 +1,0 @@
-export {}
-
-declare global {
-  const CLIENT_ID: string
-  const CLIENT_SECRET: string
-  const LOCAL: boolean
-  const authTokens: KVNamespace
-}
